refactor: keep workspace overlay under complexity budget
